pyproject.toml
[tool.poetry]
name = "bug-tracker"
version = "0.1.0"
description = ""
authors = ["Joao Gabriel <joao.gabriel156@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.1"
httpx = "^0.27.0"
alembic = "^1.13.2"

[tool.ruff]
line-length = 79
extend-exclude = ['migratios']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'TCH']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev bug_tracker/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=bug_tracker --vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
